OHRM5X-2486: migrated cypress tests
