+++ src/test/functional/.yarnrc.yml
@@ -1,3 +1,7 @@
+compressionLevel: mixed
+
+enableGlobalCache: false
+
 nodeLinker: node-modules
 
 yarnPath: .yarn/releases/yarn-4.1.0.cjs
