--- src/test/functional/cypress.json
@@ -1,7 +1,0 @@
-{
-  "baseUrl": "http://php80/orangehrm/web/index.php",
-  "defaultCommandTimeout": 6000,
-  "video": false,
-  "scrollBehavior": "top",
-  "experimentalInteractiveRunEvents": true
-}
