+++ src/test/functional/cypress/e2e/integration/feature/admin/jobCategories.cy.js
