+++ src/test/functional/cypress/e2e/integration/feature/leave/leaveTypes.cy.js
